export CardSetupResult
