+++ projects/ngx-stripe/src/lib/interfaces/payment-intent.ts
@@ -147,28 +147,3 @@ export interface PaymentMethodResult {
   paymentMethod?: PaymentMethod;
   error?: Error;
 }
-
-export interface SetupIntent {
-  id: string;
-  object: string;
-  client_secret: string;
-  created: Date;
-  description: string;
-  livemode: boolean;
-  next_action: {
-    redirect_to_url: {
-      return_url: string;
-      url: string;
-    }
-    type: string;
-    use_stripe_sdk: any;
-  };
-  payment_method: string;
-  status: string;
-  usage: string;
-}
-
-export interface CardSetupResult {
-  setupIntent?: SetupIntent;
-  error?: Error;
-}
